Shitloads
Just the highlights:
Fixed count-in movement
Greatly updated CreateController, made CreateTrack(s), allowed
scrolling while creating (a and d buttons).
Reduced the interaction options while Creating for testing purposes,
and because ctrl apparently isn't registered on Mac.
Fixed MaxPatch so that it can play the same note at one time (reducing
it to two voices in the process). Also changed sounds, so that
fail-sound is obvious to non-musicians (German).
Is implementing HighScore functionality. Got it to save formatted stuff
in a .txt, and read it again. Need to provide a viewer for it!

TODO:
Runner graphics
HiScore - GUI for inputting your score, and G for showing hiscores.
Implement proper visual design
Allow for reduced difficulty (it might be too difficult, unless we very
clearly explain what they are supposed to do)
Converted NoteLineView's calculateNoteHeight and calculateNoteIndex to
use global rather than local coordinates.
Introduced stopFlag to PlayThread. Should be safe?